LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/list-style-image-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: CSSKeywordValue
CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.

Harness Error (FAIL), message = ReferenceError: Can't find variable: CSSKeywordValue
